Install scc headers is a custom directory
It is going to be useful to have the standard headers in a known
directory which will not cause conflicts with the system headers.

+++ libc/include/i386-sysv/stdlib.h
@@ -3,7 +3,7 @@
 #define _STDLIB_H
 
 #ifndef _SIZET
-typedef unsigned size_t;
+typedef unsigned long size_t;
 #define _SIZET
 #endif
